use common WebViewHelper class
