+++ src/Controls/src/Core/WebView/WebView.cs
@@ -4,11 +4,11 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Net;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.Maui.Controls.Internals;
 
 using Microsoft.Maui.Devices;
+using Microsoft.Maui.Handlers;
 
 namespace Microsoft.Maui.Controls
 {
@@ -290,25 +290,7 @@ namespace Microsoft.Maui.Controls
 			return _platformConfigurationRegistry.Value.On<T>();
 		}
 
-		internal static string EscapeJsString(string js)
-		{
-			if (js == null)
-				return null;
-
-			if (!js.Contains('\'', StringComparison.Ordinal))
-				return js;
-
-#if NET6_0_OR_GREATER
-			return EscapeJsStringRegex().Replace(js, m =>
-#else
-			return Regex.Replace(js, @"(\\*)'", m =>
-#endif
-			{
-				int count = m.Groups[1].Value.Length;
-				// Replace with doubled backslashes plus one extra backslash, then the quote.
-				return new string('\\', (count * 2) + 1) + "'";
-			});
-		}
+		internal static string EscapeJsString(string js) => WebViewHelper.EscapeJsString(js);
 
 		/// <inheritdoc/>
 		IWebViewSource IWebView.Source => Source;
@@ -375,10 +357,5 @@ namespace Microsoft.Maui.Controls
 			var debugText = DebuggerDisplayHelpers.GetDebugText(nameof(Source), Source);
 			return $"{base.GetDebuggerDisplay()}, {debugText}";
 		}
-
-#if NET6_0_OR_GREATER
-		[GeneratedRegex(@"(\\*)'")]
-		private static partial Regex EscapeJsStringRegex();
-#endif
 	}
 }
